+++ cli/tests/test_status_command.rs
@@ -75,16 +75,29 @@ fn test_status_ignored_gitignore() {
     test_env.run_jj_in(".", ["git", "init", "repo"]).success();
     let work_dir = test_env.work_dir("repo");
 
-    let untracked_dir = work_dir.create_dir("untracked");
+    let untracked_dir = work_dir.create_dir("untracked_dir");
     untracked_dir.write_file("inside_untracked", "test");
     untracked_dir.write_file(".gitignore", "!inside_untracked\n");
-    work_dir.write_file(".gitignore", "untracked/\n!dummy\n");
+    work_dir.write_file("untracked_file", "test");
+    work_dir.write_file(".gitignore", "untracked_dir/\nuntracked_file\n!dummy\n");
 
     let output = work_dir.run_jj(["status"]);
     insta::assert_snapshot!(output, @r"
     Working copy changes:
     A .gitignore
-    Working copy  (@) : qpvuntsm 32bad97e (no description set)
+    Working copy  (@) : qpvuntsm af6b0e75 (no description set)
+    Parent commit (@-): zzzzzzzz 00000000 (empty) (no description set)
+    [EOF]
+    ");
+
+    let output = work_dir.run_jj(["status", "--ignored"]);
+    insta::assert_snapshot!(output, @r"
+    Working copy changes:
+    A .gitignore
+    Ignored paths:
+    ! untracked_dir/
+    ! untracked_file
+    Working copy  (@) : qpvuntsm af6b0e75 (no description set)
     Parent commit (@-): zzzzzzzz 00000000 (empty) (no description set)
     [EOF]
     ");